# backend/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "WebApi",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:4600",
      "description": ""
    }
  ],
  "paths": {
...
    "/api/Dates": {
      "get": {
        "tags": [
          "Dates"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DateDtoIEnumerableResponseSucceded"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DateDtoIEnumerableResponseSucceded"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DateDtoIEnumerableResponseSucceded"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Dates"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DateCreateDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DateCreateDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DateCreateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Dates/{id}": {
      "get": {
        "tags": [
          "Dates"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DateDtoResponseSucceded"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DateDtoResponseSucceded"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DateDtoResponseSucceded"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Dates"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DateDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DateDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Dates"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DateCreateDto": {
        "type": "object",
        "properties": {
          "time": {
            "type": "string",
            "format": "date-time"
          },
          "serviceId": {
            "type": "integer",
            "format": "int32"
          },
          "petId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "DateDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "time": {
            "type": "string",
            "format": "date-time"
          },
          "serviceId": {
            "type": "integer",
            "format": "int32"
          },
          "petId": {
            "type": "integer",
            "format": "int32"
          },
          "stateDate": {
            "$ref": "#/components/schemas/DateState"
          },
          "stateName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DateDtoIEnumerableResponseSucceded": {
        "type": "object",
        "properties": {
          "succeded": {
            "type": "boolean"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DateDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DateDtoResponseSucceded": {
        "type": "object",
        "properties": {
          "succeded": {
            "type": "boolean"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "data": {
            "$ref": "#/components/schemas/DateDto"
          }
        },
        "additionalProperties": false
      },
...
      "DateState": {
        "enum": [
          1,
          2,
          3
        ],
        "type": "integer",
        "format": "int32"
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Token Authorization Header using Bearer Scheme",
        "scheme": "Bearer",
        "bearerFormat": "Token"
      }
    }
  },
  "security": [
    {
      "Bearer": []
    }
  ]
}